Revert "Merge pull request #103 from ihiroky/feature/command-palette-20210526"
This reverts commit b771a99bd8fc6394204ada02fa5dd3697914bb65, reversing
changes made to f62524fda1551a659a4823128eea907df99d1f0a.
