+++ packages/desktop/src/main/index.tsx
@@ -8,7 +8,6 @@ declare global {
   interface Window {
     main: {
       request(): Promise<SettingsV1>
-      onCommandExecuted(command: string): Promise<void>
     }
   }
 }
@@ -23,7 +22,6 @@ window.main.request().then((settings: SettingsV1): void => {
           hash={createHash(settings.general.password)}
           duration={settings.general.duration * 1000}
           watermark={settings.watermark}
-          onCommandExecuted={window.main.onCommandExecuted}
         />
       </React.StrictMode>
     )
